+++ src/bind.c
@@ -1,4 +1,5 @@
 #include "vctrs.h"
+#include "slice-assign.h"
 #include "type-data-frame.h"
 #include "utils.h"
 
@@ -145,6 +146,9 @@ static SEXP vec_rbind(SEXP xs, SEXP ptype, SEXP names_to, struct name_repair_opt
   // Compact sequences use 0-based counters
   R_len_t counter = 0;
 
+  const struct vec_assign_opts bind_assign_opts =
+    new_vec_assign_opts(true, args_empty, args_empty);
+
   for (R_len_t i = 0; i < n; ++i) {
     R_len_t size = ns[i];
     if (!size) {
@@ -154,7 +158,7 @@ static SEXP vec_rbind(SEXP xs, SEXP ptype, SEXP names_to, struct name_repair_opt
 
     SEXP tbl = PROTECT(vec_cast(x, ptype, args_empty, args_empty));
     init_compact_seq(idx_ptr, counter, size, true);
-    out = df_assign(out, idx, tbl);
+    out = df_assign(out, idx, tbl, &bind_assign_opts);
     REPROTECT(out, out_pi);
 
     if (has_rownames) {