+++ src/vctrs.h
@@ -371,7 +371,7 @@ SEXP vec_slice(SEXP x, SEXP subscript);
 SEXP vec_slice_impl(SEXP x, SEXP subscript);
 SEXP vec_chop(SEXP x, SEXP indices);
 SEXP vec_slice_shaped(enum vctrs_type type, SEXP x, SEXP index);
-SEXP vec_assign(SEXP x, SEXP index, SEXP value, struct vctrs_arg* x_arg, struct vctrs_arg* value_arg);
+SEXP vec_assign(SEXP x, SEXP index, SEXP value);
 SEXP vec_proxy_assign(SEXP proxy, SEXP index, SEXP value);
 bool vec_requires_fallback(SEXP x, struct vctrs_proxy_info info);
 SEXP vec_init(SEXP x, R_len_t n);
@@ -427,10 +427,6 @@ R_len_t df_raw_size(SEXP x);
 R_len_t df_raw_size_from_list(SEXP x);
 SEXP vec_bare_df_restore(SEXP x, SEXP to, SEXP n);
 SEXP vec_df_restore(SEXP x, SEXP to, SEXP n);
-
-SEXP chr_assign(SEXP out, SEXP index, SEXP value);
-SEXP list_assign(SEXP out, SEXP index, SEXP value);
-SEXP df_assign(SEXP out, SEXP index, SEXP value);
 
 // equal_object() never propagates missingness, so
 // it can return a `bool`