+++ tests/testthat/test-slice-assign.R
@@ -461,3 +461,80 @@ test_that("slice and assign have informative errors", {
     vec_assign(1:2, 1L, 1:2, value_arg = "bar")
   })
 })
+
+test_that("names are not assigned by default", {
+  vec_x <- set_names(1:3, letters[1:3])
+  vec_y <- c(FOO = 4L)
+  vec_out <- c(a = 1L, b = 4L, c = 3L)
+  expect_identical(
+    vec_assign(vec_x, 2, vec_y),
+    vec_out
+  )
+
+  df_x <- new_data_frame(list(x = 1:3), row.names = letters[1:3])
+  df_y <- new_data_frame(list(x = 4L), row.names = "FOO")
+  df_out <- new_data_frame(list(x = c(1L, 4L, 3L)), row.names = letters[1:3])
+  expect_identical(
+    vec_assign(df_x, 2, df_y),
+    df_out
+  )
+
+  mat_x <- matrix(1:3, 3, dimnames = list(letters[1:3]))
+  mat_y <- matrix(4L, 1, dimnames = list("FOO"))
+  mat_out <- matrix(c(1L, 4L, 3L), dimnames = list(letters[1:3]))
+  expect_identical(
+    vec_assign(mat_x, 2, mat_y),
+    mat_out
+  )
+
+  nested_x <- new_data_frame(list(df = df_x, mat = mat_x, vec = vec_x), row.names = c("foo", "bar", "baz"))
+  nested_y <- new_data_frame(list(df = df_y, mat = mat_y, vec = vec_y), row.names = c("quux"))
+  nested_out <- new_data_frame(list(df = df_out, mat = mat_out, vec = vec_out), row.names = c("foo", "bar", "baz"))
+  expect_identical(
+    vec_assign(nested_x, 2, nested_y),
+    nested_out
+  )
+})
+
+test_that("can optionally assign names", {
+  vec_x <- set_names(1:3, letters[1:3])
+  vec_y <- c(FOO = 4L)
+  vec_out <- c(a = 1L, FOO = 4L, c = 3L)
+  expect_identical(
+    vec_assign_params(vec_x, 2, vec_y, assign_names = TRUE),
+    vec_out
+  )
+
+  oo_x <- set_names(as.POSIXlt(c("2020-01-01", "2020-01-02", "2020-01-03")), letters[1:3])
+  oo_y <- as.POSIXlt(c(FOO = "2020-01-04"))
+  oo_out <- as.POSIXlt(c(a = "2020-01-01", FOO = "2020-01-04", c = "2020-01-03"))
+  expect_identical(
+    vec_assign_params(oo_x, 2, oo_y, assign_names = TRUE),
+    oo_out
+  )
+
+  df_x <- new_data_frame(list(x = 1:3), row.names = letters[1:3])
+  df_y <- new_data_frame(list(x = 4L), row.names = "FOO")
+  df_out <- new_data_frame(list(x = c(1L, 4L, 3L)), row.names = c("a", "FOO", "c"))
+  expect_identical(
+    vec_assign_params(df_x, 2, df_y, assign_names = TRUE),
+    df_out
+  )
+
+  mat_x <- matrix(1:3, 3, dimnames = list(letters[1:3]))
+  mat_y <- matrix(4L, 1, dimnames = list("FOO"))
+  mat_out <- matrix(c(1L, 4L, 3L), dimnames = list(c("a", "FOO", "c")))
+  expect_identical(
+    vec_assign_params(mat_x, 2, mat_y, assign_names = TRUE),
+    mat_out
+  )
+
+  nested_x <- new_data_frame(list(df = df_x, mat = mat_x, vec = vec_x, oo = oo_x), row.names = c("foo", "bar", "baz"))
+  nested_y <- new_data_frame(list(df = df_y, mat = mat_y, vec = vec_y, oo = oo_y), row.names = c("quux"))
+  nested_out <- new_data_frame(list(df = df_out, mat = mat_out, vec = vec_out, oo = oo_out), row.names = c("foo", "quux", "baz"))
+
+  expect_identical(
+    vec_assign_params(nested_x, 2, nested_y, assign_names = TRUE),
+    nested_out
+  )
+})